+++ examples/JsonApi/Resource/AuthorResourceTransformer.php
@@ -21,6 +21,7 @@ class AuthorResourceTransformer extends AbstractResourceTransformer
      */
     public function getId($resource)
     {
+        if (isset($resource["id"]) === false) { print_r($resource); exit(); }
         return $resource["id"];
     }
 
@@ -35,10 +36,9 @@ class AuthorResourceTransformer extends AbstractResourceTransformer
 
     /**
      * @param mixed $resource
-     * @param string $relationshipPath
      * @return \WoohooLabs\Yin\JsonApi\Schema\Links|null
      */
-    public function getLinks($resource, $relationshipPath)
+    public function getLinks($resource)
     {
         return null;
     }
@@ -58,9 +58,10 @@ class AuthorResourceTransformer extends AbstractResourceTransformer
 
     /**
      * @param mixed $resource
+     * @param string $baseRelationshipPath
      * @return \WoohooLabs\Yin\JsonApi\Schema\Relationships|null
      */
-    public function getRelationships($resource)
+    public function getRelationships($resource, $baseRelationshipPath)
     {
         return null;
     }