+++ examples/JsonApi/Resource/ContactResourceTransformer.php
@@ -35,17 +35,16 @@ class ContactResourceTransformer extends AbstractResourceTransformer
 
     /**
      * @param mixed $resource
-     * @param string $relationshipPath
      * @return \WoohooLabs\Yin\JsonApi\Schema\Links|null
      */
-    public function getLinks($resource, $relationshipPath)
+    public function getLinks($resource)
     {
         return null;
     }
 
     /**
      * @param mixed $resource
-     * @return \WoohooLabs\Yin\JsonApi\Schema\Attributes
+     * @return \WoohooLabs\Yin\JsonApi\Schema\Attributes|null
      */
     public function getAttributes($resource)
     {
@@ -58,9 +57,10 @@ class ContactResourceTransformer extends AbstractResourceTransformer
 
     /**
      * @param mixed $resource
-     * @return \WoohooLabs\Yin\JsonApi\Schema\Relationships
+     * @param string $baseRelationshipPath
+     * @return \WoohooLabs\Yin\JsonApi\Schema\Relationships|null
      */
-    public function getRelationships($resource)
+    public function getRelationships($resource, $baseRelationshipPath)
     {
         return null;
     }